modified Signin return

+++ public/js/mandal.js
@@ -13,9 +13,28 @@ window.onload = async ()=>{
       title : ''
     },
     methods:{
-      goDown : (event)=>{
-        localStorage.setItem('middleNum', 1);
-        location.replace('/downmandal');
+      goDown : async (event)=>{
+        if(mandal1.isActive){
+          localStorage.setItem('middleNum', 1);
+          location.replace('/downmandal');
+        }
+        let mtitle = prompt("2단계 만다라트의 제목을 입력하세요!")
+        if(mtitle === null) alert('다시 입력해주세요!')
+        else if(mtitle != ''){
+          let result = await axios.post('/setMiddle', {
+            title : mtitle,
+            order : 0
+          })
+          if( result.status == 200){
+            localStorage.setItem('middleNum', 1);
+            location.replace('/downmandal');
+          }
+          else {
+            alert('에러!')
+          }
+        }
+        else
+          alert('다시 입력해주세요!')
       }
     }
   })
@@ -26,9 +45,28 @@ window.onload = async ()=>{
       title : ''
     },
     methods:{
-      goDown : (event)=>{
-        localStorage.setItem('middleNum', 2);
-        location.replace('/downmandal');
+      goDown : async (event)=>{
+        if(mandal2.isActive){
+          localStorage.setItem('middleNum', 1);
+          location.replace('/downmandal');
+        }
+        let mtitle = prompt("2단계 만다라트의 제목을 입력하세요!")
+        if(mtitle === null) alert('다시 입력해주세요!')
+        else if(mtitle != ''){
+          let result = await axios.post('/setMiddle', {
+            title : mtitle,
+            order : 1
+          })
+          if( result.status == 200){
+            localStorage.setItem('middleNum', 1);
+            location.replace('/downmandal');
+          }
+          else {
+            alert('에러!')
+          }
+        }
+        else
+          alert('다시 입력해주세요!')
       }
     }
   })
@@ -39,9 +77,28 @@ window.onload = async ()=>{
       title : ''
     },
     methods:{
-      goDown : (event)=>{
-        localStorage.setItem('middleNum', 3);
-        location.replace('/downmandal');
+      goDown : async (event)=>{
+        if(mandal3.isActive){
+          localStorage.setItem('middleNum', 1);
+          location.replace('/downmandal');
+        }
+        let mtitle = prompt("2단계 만다라트의 제목을 입력하세요!")
+        if(mtitle === null) alert('다시 입력해주세요!')
+        else if(mtitle != ''){
+          let result = await axios.post('/setMiddle', {
+            title : mtitle,
+            order : 2
+          })
+          if( result.status == 200){
+            localStorage.setItem('middleNum', 1);
+            location.replace('/downmandal');
+          }
+          else {
+            alert('에러!')
+          }
+        }
+        else
+          alert('다시 입력해주세요!')
       }
     }
   })
@@ -52,9 +109,28 @@ window.onload = async ()=>{
       title : ''
     },
     methods:{
-      goDown : (event)=>{
-        localStorage.setItem('middleNum', 4);
-        location.replace('/downmandal');
+      goDown : async (event)=>{
+        if(mandal4.isActive){
+          localStorage.setItem('middleNum', 1);
+          location.replace('/downmandal');
+        }
+        let mtitle = prompt("2단계 만다라트의 제목을 입력하세요!")
+        if(mtitle === null) alert('다시 입력해주세요!')
+        else if(mtitle != ''){
+          let result = await axios.post('/setMiddle', {
+            title : mtitle,
+            order : 3
+          })
+          if( result.status == 200){
+            localStorage.setItem('middleNum', 1);
+            location.replace('/downmandal');
+          }
+          else {
+            alert('에러!')
+          }
+        }
+        else
+          alert('다시 입력해주세요!')
       }
     }
   })
@@ -65,9 +141,28 @@ window.onload = async ()=>{
       title : ''
     },
     methods:{
-      goDown : (event)=>{
-        localStorage.setItem('middleNum', 5);
-        location.replace('/downmandal');
+      goDown : async (event)=>{
+        if(mandal5.isActive){
+          localStorage.setItem('middleNum', 1);
+          location.replace('/downmandal');
+        }
+        let mtitle = prompt("2단계 만다라트의 제목을 입력하세요!")
+        if(mtitle === null) alert('다시 입력해주세요!')
+        else if(mtitle != ''){
+          let result = await axios.post('/setMiddle', {
+            title : mtitle,
+            order : 4
+          })
+          if( result.status == 200){
+            localStorage.setItem('middleNum', 1);
+            location.replace('/downmandal');
+          }
+          else {
+            alert('에러!')
+          }
+        }
+        else
+          alert('다시 입력해주세요!')
       }
     }
   })
@@ -78,9 +173,28 @@ window.onload = async ()=>{
       title : ''
     },
     methods:{
-      goDown : (event)=>{
-        localStorage.setItem('middleNum', 6);
-        location.replace('/downmandal');
+      goDown : async (event)=>{
+        if(mandal6.isActive){
+          localStorage.setItem('middleNum', 1);
+          location.replace('/downmandal');
+        }
+        let mtitle = prompt("2단계 만다라트의 제목을 입력하세요!")
+        if(mtitle === null) alert('다시 입력해주세요!')
+        else if(mtitle != ''){
+          let result = await axios.post('/setMiddle', {
+            title : mtitle,
+            order : 5
+          })
+          if( result.status == 200){
+            localStorage.setItem('middleNum', 1);
+            location.replace('/downmandal');
+          }
+          else {
+            alert('에러!')
+          }
+        }
+        else
+          alert('다시 입력해주세요!')
       }
     }
   })
@@ -91,9 +205,28 @@ window.onload = async ()=>{
       title : ''
     },
     methods:{
-      goDown : (event)=>{
-        localStorage.setItem('middleNum', 7);
-        location.replace('/downmandal');
+      goDown : async (event)=>{
+        if(mandal7.isActive){
+          localStorage.setItem('middleNum', 1);
+          location.replace('/downmandal');
+        }
+        let mtitle = prompt("2단계 만다라트의 제목을 입력하세요!")
+        if(mtitle === null) alert('다시 입력해주세요!')
+        else if(mtitle != ''){
+          let result = await axios.post('/setMiddle', {
+            title : mtitle,
+            order : 6
+          })
+          if( result.status == 200){
+            localStorage.setItem('middleNum', 1);
+            location.replace('/downmandal');
+          }
+          else {
+            alert('에러!')
+          }
+        }
+        else
+          alert('다시 입력해주세요!')
       }
     }
   })
@@ -104,9 +237,28 @@ window.onload = async ()=>{
       title : ''
     },
     methods:{
-      goDown : (event)=>{
-        localStorage.setItem('middleNum', 8);
-        location.replace('/downmandal');
+      goDown : async (event)=>{
+        if(mandal8.isActive){
+          localStorage.setItem('middleNum', 1);
+          location.replace('/downmandal');
+        }
+        let mtitle = prompt("2단계 만다라트의 제목을 입력하세요!")
+        if(mtitle === null) alert('다시 입력해주세요!')
+        else if(mtitle != ''){
+          let result = await axios.post('/setMiddle', {
+            title : mtitle,
+            order : 7
+          })
+          if( result.status == 200){
+            localStorage.setItem('middleNum', 1);
+            location.replace('/downmandal');
+          }
+          else {
+            alert('에러!')
+          }
+        }
+        else
+          alert('다시 입력해주세요!')
       }
     }
   })
@@ -121,23 +273,15 @@ window.onload = async ()=>{
     location.replace('/login');
   }
   else if(result.status == 200){
-    mandal1.title = result.data.re.mandal[0].middleTitle;
-    if(result.data.re.mandal[0].middleTitle != '') mandal1.isActive = true;
-    mandal2.title = result.data.re.mandal[1].middleTitle;
-    if(result.data.re.mandal[1].middleTitle != '') mandal2.isActive = true;
-    mandal3.title = result.data.re.mandal[2].middleTitle;
-    if(result.data.re.mandal[2].middleTitle != '') mandal3.isActive = true;
-    mandal4.title = result.data.re.mandal[3].middleTitle;
-    if(result.data.re.mandal[3].middleTitle != '') mandal4.isActive = true;
-    mandal5.title = result.data.re.mandal[4].middleTitle;
-    if(result.data.re.mandal[4].middleTitle != '') mandal5.isActive = true;
-    mandal6.title = result.data.re.mandal[5].middleTitle;
-    if(result.data.re.mandal[5].middleTitle != '') mandal6.isActive = true;
-    mandal7.title = result.data.re.mandal[6].middleTitle;
-    if(result.data.re.mandal[6].middleTitle != '') mandal7.isActive = true;
-    mandal8.title = result.data.re.mandal[7].middleTitle;
-    if(result.data.re.mandal[7].middleTitle != '') mandal8.isActive = true;
     mandalTitle.title = result.data.re.title;
     mandalTitle.achievement = result.data.re.achievement;
+    if(result.data.re.mandal[0].middleTitle != '') mandal1.isActive = true;
+    if(result.data.re.mandal[1].middleTitle != '') mandal2.isActive = true;
+    if(result.data.re.mandal[2].middleTitle != '') mandal3.isActive = true;
+    if(result.data.re.mandal[3].middleTitle != '') mandal4.isActive = true;
+    if(result.data.re.mandal[4].middleTitle != '') mandal5.isActive = true;
+    if(result.data.re.mandal[5].middleTitle != '') mandal6.isActive = true;
+    if(result.data.re.mandal[6].middleTitle != '') mandal7.isActive = true;
+    if(result.data.re.mandal[7].middleTitle != '') mandal8.isActive = true;
   }
 }
